Additional merkle tests
