+++ merkle/nodecontainer.go
@@ -1,6 +1,7 @@
 package merkle
 
 import (
+	"bytes"
 	"encoding/hex"
 	"errors"
 	"fmt"
@@ -35,6 +36,7 @@ type NodeContainer interface {
 	getNodeEmbeddedPath() string
 
 	//updateChildPointer(prefix string, child NodeContainer)
+	print(userDb *leveldb.DB, treeDb *leveldb.DB) string
 }
 
 type nodeContainerImp struct {
@@ -320,4 +322,33 @@ func (n *nodeContainerImp) marshal() ([]byte, error) {
 	default:
 		return nil, errors.New(fmt.Sprintf("unexpcted node type %d", n.nodeType))
 	}
+}
+
+// depth-first-search print tree rooted at node n
+// note - this will load the whole tree into memory
+
+func (n *nodeContainerImp) print(userDb *leveldb.DB, treeDb *leveldb.DB) string {
+
+	buffer := bytes.Buffer{}
+
+	n.loadChildren(treeDb)
+
+	switch n.nodeType {
+	case pb.NodeType_branch:
+		for _, v := range n.children {
+			buffer.WriteString(v.print(userDb, treeDb))
+		}
+		buffer.WriteString(n.getBranchNode().print())
+
+	case pb.NodeType_leaf:
+		buffer.WriteString(n.getLeafNode().print())
+
+	case pb.NodeType_extension:
+		for _, v := range n.children {
+			buffer.WriteString(v.print(userDb, treeDb))
+		}
+		buffer.WriteString(n.getExtNode().print())
+	}
+
+	return buffer.String()
 }
